API

Bolt.customion

Bolt.peebles_Xₑ

Bolt.saha_peebles_recombination

Bolt.customion — Method

Convenience function to create an ionisation history from some tables

Bolt.peebles_Xₑ — Method

peebles_Xₑ(par, Xₑ₀, x_start, x_end)

Solve the Peebles equation over a span of scale factors, and then construct an interpolator mapping scale factor to the resulting ionization fraction.

Arguments:

par: cosmological parameters

Xₑ₀: initial ionization fraction

x_start: scale factor to begin integration

x_end: scale factor to end integration

Returns:

generic function: interpolator for Xₑ(x)

Bolt.saha_peebles_recombination — Method

saha_peebles_recombination(par::AbstractCosmoParams)

Utility function for generating a decent approximation to Xₑ in ΛCDM recombination, using the Saha equation until z=1587.4 and then the Peebles equation for the rest.